third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-size-adjust-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: getComputedStyle().fontSizeAdjust</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-size-adjust-prop">
<meta name="assert" content="font-size-adjust computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('font-size-adjust', 'none');
test_computed_value('font-size-adjust', '0.5');
test_computed_value('font-size-adjust', 'ex 0.5', '0.5');  // default basis 'ex' omitted from serialization
test_computed_value('font-size-adjust', 'cap 0.8');
test_computed_value('font-size-adjust', 'ch 0.4');
test_computed_value('font-size-adjust', 'ic 0.9');
</script>
</body>
</html>
