third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-size-adjust-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property font-size-adjust value 'ic-height from-font'
  assert_equals: expected "ic-height 1" but got "none"
Harness: the test ran to completion.
